Split out holidays into separate controller
